requirements.txt
fastapi
uvicorn
pillow
python-multipart
ultralytics
opencv-python-headless